renamed dev_atheloph.py to ateloph.py
